fix: refactor ProjectPiechart to funtional component
