+++ src/components/Reports/ProjectReport/ProjectPieChart/ProjectPieChart.jsx
@@ -1,0 +1,92 @@
+import React, { useState } from 'react';
+import { PieChart, Pie, Sector, ResponsiveContainer } from 'recharts';
+
+
+const generateRandomHexColor = () => {
+
+  const randomColor = Math.floor(Math.random() * 16777215).toString(16);
+  const hexColor = "#" + "0".repeat(6 - randomColor.length) + randomColor;
+  return hexColor;
+}
+
+const renderActiveShape = (props) => {
+  const hexColor = generateRandomHexColor()
+  const RADIAN = Math.PI / 180;
+  const { cx, cy, midAngle, innerRadius, outerRadius, startAngle, endAngle, fill, payload, percent, value } = props;
+  const sin = Math.sin(-RADIAN * midAngle);
+  const cos = Math.cos(-RADIAN * midAngle);
+  const sx = cx + (outerRadius + 10) * cos;
+  const sy = cy + (outerRadius + 10) * sin;
+  const mx = cx + (outerRadius + 30) * cos;
+  const my = cy + (outerRadius + 30) * sin;
+  const ex = mx + (cos >= 0 ? 1 : -1) * 22;
+  const ey = my;
+  const textAnchor = cos >= 0 ? 'start' : 'end';
+
+
+  return (
+    <g>
+      <text x={cx} y={cy} dy={8} textAnchor="middle" fill={fill}>
+        {payload.lastName.substring(0, 5)} {payload.value.toFixed(1)} of {payload.totalHoursCalculated.toFixed(1)}hrs
+      </text>
+      <Sector
+        cx={cx}
+        cy={cy}
+        innerRadius={innerRadius}
+        outerRadius={outerRadius}
+        startAngle={startAngle}
+        endAngle={endAngle}
+        fill={hexColor}
+      />
+      <Sector
+        cx={cx}
+        cy={cy}
+        startAngle={startAngle}
+        endAngle={endAngle}
+        innerRadius={outerRadius + 6}
+        outerRadius={outerRadius + 10}
+        fill={hexColor}
+      />
+      <path d={`M${sx},${sy}L${mx},${my}L${ex},${ey}`} stroke={fill} fill="none" />
+      <circle cx={ex} cy={ey} r={2} fill={fill} stroke="none" />
+      <text x={ex + (cos >= 0 ? 1 : -1) * 12} y={ey} textAnchor={textAnchor} fill="#333">{`${payload.name.substring(0, 14)}`}</text>
+      <text x={ex + (cos >= 0 ? 1 : -1) * 12} y={ey} dy={18} textAnchor={textAnchor} fill="#999">
+        {`${value.toFixed(2)} Hours (${(percent * 100).toFixed(2)}%)`}
+      </text>
+    </g>
+  );
+};
+
+export function ProjectPieChart  ({ userData, windowSize }) {
+  const [activeIndex, setActiveIndex] = useState(0);
+
+  const onPieEnter = (_, index) => {
+    setActiveIndex(index);
+  };
+
+  let circleSize = 30;
+  if (windowSize <= 1280) {
+    circleSize = windowSize / 10 * 0.5;
+  }
+
+  return (
+    <>
+      <ResponsiveContainer maxWidth={640} maxHeight={640} minWidth={350} minHeight={350}>
+        <PieChart>
+          <Pie
+            activeIndex={activeIndex}
+            activeShape={renderActiveShape}
+            data={userData}
+            cx="50%"
+            cy="50%"
+            innerRadius={60 + circleSize}
+            outerRadius={120 + circleSize}
+            fill="#8884d8"
+            dataKey="value"
+            onMouseEnter={onPieEnter}
+            />
+          </PieChart>
+        </ResponsiveContainer>
+      </>
+  );
+}
