--- src/components/Reports/ProjectReport/WbsPiechart/data.js
@@ -1,32 +1,0 @@
-export const data = [
-  {
-    "id": "make",
-    "label": "make",
-    "value": 132,
-    "color": "hsl(289, 70%, 50%)"
-  },
-  {
-    "id": "haskell",
-    "label": "haskell",
-    "value": 122,
-    "color": "hsl(23, 70%, 50%)"
-  },
-  {
-    "id": "sass",
-    "label": "sass",
-    "value": 493,
-    "color": "hsl(106, 70%, 50%)"
-  },
-  {
-    "id": "erlang",
-    "label": "erlang",
-    "value": 317,
-    "color": "hsl(350, 70%, 50%)"
-  },
-  {
-    "id": "scala",
-    "label": "scala",
-    "value": 379,
-    "color": "hsl(57, 70%, 50%)"
-  }
-]
